migrati ulang

+++ database/migrations/2020_04_03_075721_create_issuings_table.php
@@ -16,7 +16,7 @@ class CreateIssuingsTable extends Migration
         Schema::create('issuings', function (Blueprint $table) {
             $table->increments('id');
             $table->String('date');
-            $table->unsignedInteger('customer_id');
+            $table->unsignedBigInteger('customer_id');
             $table->string('bayar');
             $table->string('kembali');
             $table->string('grandtotal');
